DEV-130 reorganize code
Moved code out of __main__ and __init__.
