+++ gdc_maf_tool/log.py
@@ -1,0 +1,15 @@
+import sys
+
+import cdislogging
+
+logger = cdislogging.get_logger("gdc-maf-tool", log_level="info")
+
+
+def fatal(message: str):
+    """
+    Report the error and exit the program.
+
+    Python's logger.fatal does not exit the program.
+    """
+    logger.fatal(message)
+    sys.exit(2)
